Rename athens_integration -> e2etests
